// mage-client-web/src/lib/generated/mage/v1/room.ts
export interface RoomGetTableByIdRequest {
	sessionId: string;
	roomId: string;
	tableId: string;
}

export interface RoomGetTableByIdResponse {
	table?: TableView | undefined;
}
// ...
export const RoomGetTableByIdRequest: MessageFns<RoomGetTableByIdRequest> = {
	encode(
		message: RoomGetTableByIdRequest,
		writer: BinaryWriter = new BinaryWriter()
	): BinaryWriter {
		if (message.sessionId !== '') {
			writer.uint32(10).string(message.sessionId);
		}
		if (message.roomId !== '') {
			writer.uint32(18).string(message.roomId);
		}
		if (message.tableId !== '') {
			writer.uint32(26).string(message.tableId);
		}
		return writer;
	},

	decode(input: BinaryReader | Uint8Array, length?: number): RoomGetTableByIdRequest {
		const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
		const end = length === undefined ? reader.len : reader.pos + length;
		const message = createBaseRoomGetTableByIdRequest();
		while (reader.pos < end) {
			const tag = reader.uint32();
			switch (tag >>> 3) {
				case 1: {
					if (tag !== 10) {
						break;
					}

					message.sessionId = reader.string();
					continue;
				}
				case 2: {
					if (tag !== 18) {
						break;
					}

					message.roomId = reader.string();
					continue;
				}
				case 3: {
					if (tag !== 26) {
						break;
					}

					message.tableId = reader.string();
					continue;
				}
			}
			if ((tag & 7) === 4 || tag === 0) {
				break;
			}
			reader.skip(tag & 7);
		}
		return message;
	},

	fromJSON(object: any): RoomGetTableByIdRequest {
		return {
			sessionId: isSet(object.sessionId) ? globalThis.String(object.sessionId) : '',
			roomId: isSet(object.roomId) ? globalThis.String(object.roomId) : '',
			tableId: isSet(object.tableId) ? globalThis.String(object.tableId) : ''
		};
	},

	toJSON(message: RoomGetTableByIdRequest): unknown {
		const obj: any = {};
		if (message.sessionId !== '') {
			obj.sessionId = message.sessionId;
		}
		if (message.roomId !== '') {
			obj.roomId = message.roomId;
		}
		if (message.tableId !== '') {
			obj.tableId = message.tableId;
		}
		return obj;
	},

	create(base?: DeepPartial<RoomGetTableByIdRequest>): RoomGetTableByIdRequest {
		return RoomGetTableByIdRequest.fromPartial(base ?? {});
	},
	fromPartial(object: DeepPartial<RoomGetTableByIdRequest>): RoomGetTableByIdRequest {
		const message = createBaseRoomGetTableByIdRequest();
		message.sessionId = object.sessionId ?? '';
		message.roomId = object.roomId ?? '';
		message.tableId = object.tableId ?? '';
		return message;
	}
};

function createBaseRoomGetTableByIdResponse(): RoomGetTableByIdResponse {
	return { table: undefined };
}

export const RoomGetTableByIdResponse: MessageFns<RoomGetTableByIdResponse> = {
	encode(
		message: RoomGetTableByIdResponse,
		writer: BinaryWriter = new BinaryWriter()
	): BinaryWriter {
		if (message.table !== undefined) {
			TableView.encode(message.table, writer.uint32(10).fork()).join();
		}
		return writer;
	},

	decode(input: BinaryReader | Uint8Array, length?: number): RoomGetTableByIdResponse {
		const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
		const end = length === undefined ? reader.len : reader.pos + length;
		const message = createBaseRoomGetTableByIdResponse();
		while (reader.pos < end) {
			const tag = reader.uint32();
			switch (tag >>> 3) {
				case 1: {
					if (tag !== 10) {
						break;
					}

					message.table = TableView.decode(reader, reader.uint32());
					continue;
				}
			}
			if ((tag & 7) === 4 || tag === 0) {
				break;
			}
			reader.skip(tag & 7);
		}
		return message;
	},

	fromJSON(object: any): RoomGetTableByIdResponse {
		return { table: isSet(object.table) ? TableView.fromJSON(object.table) : undefined };
	},

	toJSON(message: RoomGetTableByIdResponse): unknown {
		const obj: any = {};
		if (message.table !== undefined) {
			obj.table = TableView.toJSON(message.table);
		}
		return obj;
	},

	create(base?: DeepPartial<RoomGetTableByIdResponse>): RoomGetTableByIdResponse {
		return RoomGetTableByIdResponse.fromPartial(base ?? {});
	},
	fromPartial(object: DeepPartial<RoomGetTableByIdResponse>): RoomGetTableByIdResponse {
		const message = createBaseRoomGetTableByIdResponse();
		message.table =
			object.table !== undefined && object.table !== null
				? TableView.fromPartial(object.table)
				: undefined;
		return message;
	}
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

type DeepPartial<T> = T extends Builtin
	? T
	: T extends globalThis.Array<infer U>
		? globalThis.Array<DeepPartial<U>>
		: T extends ReadonlyArray<infer U>
			? ReadonlyArray<DeepPartial<U>>
			: T extends {}
				? { [K in keyof T]?: DeepPartial<T[K]> }
				: Partial<T>;
// ...
interface MessageFns<T> {
	encode(message: T, writer?: BinaryWriter): BinaryWriter;
	decode(input: BinaryReader | Uint8Array, length?: number): T;
	fromJSON(object: any): T;
	toJSON(message: T): unknown;
	create(base?: DeepPartial<T>): T;
	fromPartial(object: DeepPartial<T>): T;
}